app/vmctl: remove flags provided by cli package
