+++ app/vmctl/remoteread.go
@@ -2,6 +2,7 @@ package main
 
 import (
 	"context"
+	"flag"
 	"fmt"
 	"log"
 	"sync"
@@ -11,8 +12,28 @@ import (
 	"github.com/VictoriaMetrics/VictoriaMetrics/app/vmctl/remoteread"
 	"github.com/VictoriaMetrics/VictoriaMetrics/app/vmctl/stepper"
 	"github.com/VictoriaMetrics/VictoriaMetrics/app/vmctl/vm"
+	"github.com/VictoriaMetrics/VictoriaMetrics/lib/flagutil"
 	"github.com/VictoriaMetrics/VictoriaMetrics/lib/logger"
 	"github.com/cheggaaa/pb/v3"
+)
+
+var (
+	remoteRead                 = flag.Bool("remote-read", false, "Use Prometheus remote read protocol")
+	remoteReadUseStream        = flag.Bool("remote-read-use-stream", false, "Defines whether to use SAMPLES or STREAMED_XOR_CHUNKS mode. By default is uses SAMPLES mode. See https://prometheus.io/docs/prometheus/latest/querying/remote_read_api/#streamed-chunks")
+	remoteReadConcurrency      = flag.Int("remote-read-concurrency", 1, "Number of concurrently running remote read readers")
+	remoteReadFilterTimeStart  = flagutil.NewTime("remote-read-filter-time-start", "", "The time filter in RFC3339 format to select timeseries with timestamp equal or higher than provided value. E.g. '2020-01-01T20:07:00Z'")
+	remoteReadFilterTimeEnd    = flagutil.NewTime("remote-read-filter-time-end", "", "The time filter in RFC3339 format to select timeseries with timestamp equal or lower than provided value. E.g. '2020-01-01T20:07:00Z'")
+	remoteReadFilterLabel      = flag.String("remote-read-filter-label", "__name__", "Prometheus label name to filter timeseries by. E.g. '__name__' will filter timeseries by name.")
+	remoteReadFilterLabelValue = flag.String("remote-read-filter-label-value", ".*", "Prometheus regular expression to filter label from \"remote-read-filter-label\" flag.")
+	remoteReadStepInterval     = flag.String("remote-read-step-interval", "", fmt.Sprintf("Split export data into chunks. Requires setting --%s. Valid values are %q,%q,%q,%q.", "remote-read-filter-time-start", stepper.StepMonth, stepper.StepDay, stepper.StepHour, stepper.StepMinute))
+	remoteReadSrcAddr          = flag.String("remote-read-src-addr", "", "Remote read address to perform read from.")
+	remoteReadUser             = flag.String("remote-read-user", "", "Remote read username for basic auth")
+	remoteReadPassword         = flag.String("remote-read-password", "", "Remote read password for basic auth")
+	remoteReadHTTPTimeout      = flag.Duration("remote-read-http-timeout", 0, "Timeout defines timeout for HTTP write request to remote storage")
+	remoteReadHeaders          = flag.String("remote-read-headers", "", "Optional HTTP headers to send with each request to the corresponding remote source storage \n"+
+		"For example, --remote-read-headers='My-Auth:foobar' would send 'My-Auth: foobar' HTTP header with every request to the corresponding remote source storage. \n"+
+		"Multiple headers must be delimited by '^^': --remote-read-headers='header1:value1^^header2:value2'")
+	remoteReadInsecureSkipVerify = flag.Bool("remote-read-insecure-skip-verify", false, "Whether to skip TLS certificate verification when connecting to the remote read address")
 )
 
 type remoteReadProcessor struct {
@@ -129,6 +150,18 @@ func (rrp *remoteReadProcessor) do(ctx context.Context, filter *remoteread.Filte
 
 func remoteReadImport(ctx context.Context) func([]string) {
 	return func(args []string) {
+		err := flagutil.SetFlagsFromEnvironment()
+		if err != nil {
+			logger.Fatalf("error set flags from environment variables: %s", err)
+		}
+
+		if *remoteReadSrcAddr == "" {
+			logger.Fatalf("flag --remote-read-src-addr cannot be empty")
+		}
+		if *remoteReadStepInterval == "" {
+			logger.Fatalf("flag --remote-read-step-interval cannot be empty")
+		}
+
 		rr, err := remoteread.NewClient(remoteread.Config{
 			Addr:               *remoteReadSrcAddr,
 			Username:           *remoteReadUser,
