+++ lib/flagutil/flag.go
@@ -4,6 +4,7 @@ import (
 	"flag"
 	"fmt"
 	"io"
+	"os"
 	"strings"
 )
 
@@ -17,4 +18,22 @@ func WriteFlags(w io.Writer) {
 		}
 		fmt.Fprintf(w, "-%s=%q\n", f.Name, value)
 	})
+}
+
+// SetFlagsFromEnvironment iterates over the defined flags using flag.VisitAll and
+// attempts to set their value from the environment using flag.Set.
+// For example, the function above sets -my-flag from the environment variable MY_FLAG if it exists.
+func SetFlagsFromEnvironment() error {
+	var err error
+	flag.VisitAll(func(f *flag.Flag) {
+		name := strings.ToUpper(strings.Replace(f.Name, "-", "_", -1))
+		if value, ok := os.LookupEnv(name); ok {
+			err = flag.Set(f.Name, value)
+			if err != nil {
+				err = fmt.Errorf("failed setting flag from environment: %w", err)
+			}
+		}
+	})
+
+	return err
 }
